Use most parameters from Credentials Cache key
The Credentials Cache produces cache hits based on the cache key,
but the cached data is generated from parameters that are not
always reflected in the key (e.g. `PolaricPrincipal`). As a result,
it is not evident that positive cache hits always represent exactly
the same credentials that would be produced from scratch.

This change takes care of propagating all parameter through the cache
key, except the entity, which is a bit complicated to handle and
probably needs a separate PR.